+++ CHANGELOG.md
@@ -494,3 +494,4 @@
 * Define experimental `Traversable_` typeclass
 * Define experimental `Distributibe_` typeclass
 * Define experimental `-<<-<<-` method of `Traversable_` typeclass
+* Remove `Adjoint` typeclass